+++ CHANGELOG.md
@@ -10,11 +10,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [ ] warn for superfluous prefix lengths applicable mechanisms (i.e. /32 resp. /128)
 - [ ] create dot-file that visualizes the records (?)
-- [ ] add spf domain when warning about string length exceeding 512 characters
+
+### changed
+
+- added SPF domain name in warning when exceeding 512 characters
+- seen before warning changed into multiple entries warning
+
+### added
+
+- case of multiple entries, warn about inconsistencies (if any)
+
 
 ## [v0.2.0] - 2021-11-21
 
 ### Changed
+
 - verdict output includes owner domain and contact (also in csv-output)
 - ipt logs show spf terms rather than their raw token
 - logging to stderr now shows the domain in front, so redirecting stderr to a
@@ -22,14 +32,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   that time.
 - added warning when ?all or +all is used
 
-
 ### Fixed
+
 - url for rfc7208 test suite
 - use :dns (not :ipt) when logging dns additions to the cache
 
+
 ## [v0.1.1] - 2021-11-20
+
 - Fix url for License badge
 
-
 ## [v0.1.0] - 2021-11-20
+
 - Initial public version